+++ recipai-frontend/src/screens/PromptTest/index.js
@@ -6,62 +6,21 @@ import Grid2 from '@mui/material/Unstable_Grid2/Grid2'
 import PromptTextBox from '../../components/PromptTextBox'
 import SendPromptButton from '../../components/SendPromptButton'
 
+import InitialScreenSection from './InitialScreenSection'
 import mockRequest from '../../mock/mockServer.js'
 import { clamp, lerp, pct } from '../../util/utils'
 
 
 const PromptTest = () => {
 
-    const initialPadding = 25;
 
-    const [classes, setClasses] = useState({
-        container: {
-            paddingTop: pct(initialPadding)
-        }
-    });
-
-    const [text, setText] = useState("");
     const [responseText, setResponseText] = useState("");
-
-    const [initialPosition, setInitialPosition] = useState(true)
-    const [padding, setPadding] = useState(pct(initialPadding));
-
-    const [slideInterval, setSlideInterval] = useState(null)
-    const [isIntervalDone, setIsIntervalDone] = useState(false)
-
 
     useEffect(() => {
         console.log(responseText);
     }, [responseText]);
 
-    useEffect(() => {
-        setClasses({
-            ...classes,
-            container: {
-                ...classes.container,
-                paddingTop: pct(padding)
-            }
-        })
-
-        if (padding == 0 && !isIntervalDone) {
-            setIsIntervalDone(true)
-        }
-        //console.log(padding)
-    }, [padding])
-
-    useEffect(() => {
-        if (isIntervalDone) {
-            clearInterval(slideInterval);
-            setSlideInterval(null);
-            setIsIntervalDone(false);
-        }
-    }, [isIntervalDone])
-
-    const handleInputChange = (event) => {
-        setText(event.target.value);
-    }
-
-    const handlePromptSend = useCallback(() => {
+    const handlePromptSend = useCallback((text) => {
         fetch("/api/recipe", {
             method: 'POST',
             headers: {
@@ -79,7 +38,7 @@ const PromptTest = () => {
             })
     }, [text])
 
-    const handleMockPrompt = useCallback(() => {
+    const handleMockPrompt = useCallback((text) => {
         mockRequest()
             .then(res => {
                 console.log(res)
@@ -87,68 +46,38 @@ const PromptTest = () => {
             });
     }, [text])
 
-    const handleButtonPress = useCallback(() => {
+    const sendPromptCallback = useCallback((text) => {
         if (process.env.NODE_ENV == 'development') {
             console.log("Mock response for dev")
             //TODO only when testing api 
             //handlePromptSend();
             //All other times
-            handleMockPrompt();
+            handleMockPrompt(text);
 
         }
+
         else if (process.env.NODE_ENV == 'test') {
             console.log("Test mock here?")
         }
 
         else if (process.env.NODE_ENV == 'production') {
             console.log("DO real API code here")
-            handlePromptSend();
+            handlePromptSend(text);
         }
 
         else {
             console.log("node env not found")
         }
 
-        if (initialPosition) {
-            setInitialPosition(false);
-            let startTime = (new Date()).getTime();
-            setSlideInterval(setInterval(() => {
-                let currentTime = (new Date()).getTime();
-                console.log("TIME VALUE", (currentTime - startTime) / 1000)
-                if (((currentTime - startTime) / 1000) < 0) {
-                    console.log("UH WHY IS THIS NEGATIVE", currentTime, startTime);
-                }
-                // console.log(clamp(0, 1, (currentTime - startTime) / 1000))
-                // console.log("LERP", lerp(initialPadding, 0, clamp(0, 1, 1 - (currentTime - startTime) / 1000)))
-                let lerpVal = lerp(initialPadding, 0, clamp(0, 1, 1 - (currentTime - startTime) / 1000))
-                setPadding(lerpVal);
-            }), 1000 / 10)
-        }
 
-        return () => {
-            if (slideInterval) {
-                clearInterval(slideInterval);
-                setSlideInterval(null);
-            }
-        }
-    },)
+    })
 
     return (
         <>
             <Grid2 alignItems="center" container direction="column" spacing={12}>
                 <Grid2 xs={12}>
                 </Grid2>
-                <Grid2 xs={12} style={classes.container}>
-                    <Grid2 alignItems="center" container direction="column" spacing={1}>
-                        <Grid2 xs={12}>
-                            <TextField value={text} onChange={handleInputChange} />
-                        </Grid2>
-                        <Grid2 xs={12}>
-                            <SendPromptButton handleButtonPress={handleButtonPress} />
-                        </Grid2>
-
-                    </Grid2>
-                </Grid2>
+                <InitialScreenSection sendPromptCallback={sendPromptCallback} />
                 <Grid2 xs={12}>
                     <PromptTextBox text={responseText} />
                 </Grid2>
